+++ packages/react-native-passkey-stamper/src/__tests__/util-test.ts
@@ -1,9 +1,5 @@
 import { test, expect } from "@jest/globals";
-import {
-  base64Tobase64url,
-  getChallengeFromPayload,
-  getRandomChallenge,
-} from "../util";
+import { getChallengeFromPayload, getRandomChallenge } from "../util";
 
 test("generates correct challenges from payloads", async function () {
   // You can verify that this is the right value with:
@@ -27,11 +23,4 @@ test("generates valid random challenges", async function () {
   const buffer = Buffer.from(challenge, "hex");
   // 64 hex characters should be exactly 32 bytes
   expect(buffer.length).toBe(32);
-});
-
-test("converts base64 to base64url", async function () {
-  expect(
-    // Sample credential ID returned by f-23/react-native-passkey lib
-    base64Tobase64url("XkLQIv/W4UwcSn++zaY9lrAgOj4=")
-  ).toBe("XkLQIv_W4UwcSn--zaY9lrAgOj4");
 });