+++ packages/ethers/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @turnkey/ethers
 
+## 0.19.9
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@2.7.1
+
+## 0.19.8
+
+### Patch Changes
+
+- Updated dependencies ([c3b423b], [d73725b])
+  - @turnkey/api-key-stamper@0.4.0
+  - @turnkey/http@2.7.0
+
 ## 0.19.7
 
 ### Patch Changes