+++ packages/react-native-passkey-stamper/CHANGELOG.md
@@ -1,5 +1,21 @@
 # @turnkey/react-native-passkey-stamper
 
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@2.7.1
+
+## 0.2.0
+
+### Minor Changes
+
+- Introduce `@turnkey/encoding` to consolidate utility functions
+- Updated dependencies [d73725b]
+  - @turnkey/encoding@0.1.0
+  - @turnkey/http@2.7.0
+
 ## 0.1.0
 
 Initial release