+++ packages/solana/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @turnkey/solana
 
+## 0.2.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@2.7.1
+
+## 0.2.1
+
+### Patch Changes
+
+- Updated dependencies [d73725b]
+  - @turnkey/http@2.7.0
+
 ## 0.2.0
 
 ### Minor Changes