+++ packages/viem/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @turnkey/viem
 
+## 0.4.10
+
+### Patch Changes
+
+- Updated dependencies
+  - @turnkey/http@2.7.1
+
+## 0.4.9
+
+### Patch Changes
+
+- Updated dependencies ([c3b423b], [d73725b])
+  - @turnkey/api-key-stamper@0.4.0
+  - @turnkey/http@2.7.0
+
 ## 0.4.8
 
 ### Patch Changes